404: Comment ID not found error handling and testing

+++ api/models/comments.models.js
@@ -60,6 +60,9 @@ exports.addComment = async (id, commentValues) => {
 exports.deleteCommentById = async (id) => {
 const {comment_id} = id;
 
+const checkId = await checkValueExists("comments", "comment_id", comment_id);
+if (!checkId) return Promise.reject({ status: 404, msg: "Comment ID not found" });
+
 const commentDeleted = await db.query(`
 DELETE FROM comments
 WHERE comment_id = $1
